refactor(tab-manager): inline tab-actions into .withActions() handlers
Delete tab-actions.ts and move Chrome API execution logic directly
into the workspace .withActions() mutation handlers. The file was a
single-consumer indirection layer — every execute* function was only
called from one corresponding handler in workspace.ts.

Adds nativeTabId helper alongside the existing composite ID parsers
and imports generateId for the save handler.
